+++ CHANGELOG.md
@@ -4,6 +4,9 @@
 ### Added
 - Send unscheduled alarm event
 
+### Fixed
+- Fix issue when multiple alarms scheduled for today and tomorrow. Only one was scheduled
+
 ## [1.0.0] - 2022-01-15
 
 ### Added